[feat] ViewModel을 이용한 라이브데이터 적용
