+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.Maplemate"
         tools:targetApi="31">
         <activity
-            android:name=".LogintActivity"
+            android:name=".UI.LogintActivity"
             android:exported="true">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
@@ -30,7 +30,7 @@
             android:name=".UI.TestActivity"
             android:exported="true" />
         <activity
-            android:name=".MainActivity"
+            android:name=".UI.MainActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustNothing">
         </activity>
