+++ app/src/main/res/layout/activity_logint.xml
@@ -5,7 +5,7 @@
     android:id="@+id/btn_delete"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LogintActivity">
+    tools:context=".UI.LogintActivity">
 
     <Button
         android:id="@+id/btn_login"
